[FEAT] Docker-compose
Added a docker compose to run the server. Made some necessary changes such as moving the app.py to the compiler folder and adding a requirements.txt for quick dependency fetching, also added some dependencies that the frontend project used but were not listed in the package. Removed the main.py that was legacy

Signed-off-by: elmancho0 <granadosgabriel34@gmail.com>