[FIX] Refactor parser
The parser was refactored to make more sense. The main changes are: tokens are now a class, this makes more sense than using them as an array or dictionary, methods were split so that they only do one thing or call other methods to do several things, methods and variables were given a more descriptive name. The way the parser gives errors was also changed, it now shows the line where the error is, this is something that needs to be fixed since it shows the line several times.

Signed-off-by: elmancho0 <granadosgabriel34@gmail.com>