[FEAT] Arrays
Implemented support for arrays in the semantic analyzer with size and type validation. In the SemanticAnalyzer file: The _analyze_index_op function was added to handle access to matrices and vectors. Modifications to analyze_binary_op to support assignments and validation of elements in matrices and arrays. _validate_matrix_declaration and _validate_array_declaration check types and sizes in arrays. Frontend changes in BloodCodeCompiler for execution without storing current selection of options. Modifications in CodeOptionsComponent: Removed selectedOption state to allow multiple selection. Modified the handleValueChange to immediately apply each selected option.